+++ src/main/uploader.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-lines */
 import { API } from '@sentry/core';
 import { Event, Status, Transport } from '@sentry/types';
-import { Dsn, logger, timestampWithMs } from '@sentry/utils';
+import { Dsn, forget, logger, timestampWithMs } from '@sentry/utils';
 import { basename, join } from 'path';
 
 import { supportsCrashpadOnWindows } from '../electron-version';
@@ -197,11 +197,35 @@ export class MinidumpUploader {
 
   /** Scans the Crashpad directory structure for minidump files. */
   private async _scanCrashpadFolder(): Promise<string[]> {
+    forget(this._deleteCrashpadMetadataFile());
+
     // Crashpad moves minidump files directly into the 'completed' or 'reports' folder. We can
     // load them from there, upload to the server, and then delete it.
     const dumpDirectory = join(this._crashesDirectory, this._crashpadSubDirectory);
     const files = await readDirAsync(dumpDirectory);
     return files.filter(file => file.endsWith('.dmp')).map(file => join(dumpDirectory, file));
+  }
+
+  /** Attempts to remove the metadata file so Crashpad doesn't output `failed to stat report` errors to the console */
+  private async _deleteCrashpadMetadataFile(waitMs: number = 100): Promise<void> {
+    if (waitMs > 2000) {
+      return;
+    }
+
+    const metadataPath = join(this._crashesDirectory, 'metadata');
+    try {
+      await unlinkAsync(metadataPath);
+      logger.log('Deleted Crashpad metadata file', metadataPath);
+    } catch (e) {
+      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      if (e.code && e.code == 'EBUSY') {
+        // Since Crashpad probably still has the metadata file open, we make a few attempts to delete it, backing
+        // off and waiting longer each time.
+        setTimeout(async () => {
+          await this._deleteCrashpadMetadataFile(waitMs * 2);
+        }, waitMs);
+      }
+    }
   }
 
   /** Scans the Breakpad directory structure for minidump files. */